==> llm_experiements/18_DistilBERT_scrapped_news.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import sys
from pathlib import Path

if "workding_dir" not in locals():
    workding_dir = str(Path.cwd().parent)
os.chdir(workding_dir)
sys.path.append(workding_dir)
print("working dir:", workding_dir)

from dotenv import find_dotenv, load_dotenv

found_dotenv = find_dotenv(".env")

if len(found_dotenv) == 0:
    found_dotenv = find_dotenv(".env.example")
print(f"loading env vars from: {found_dotenv}")
load_dotenv(found_dotenv, override=True)

from llm_experiements.utils import *
from datasets import load_dataset

dataset = load_dataset(
    "csv",
    data_files={
        "test": "dataset/scrapped_news.csv",
    },
)
dataset

%%time

model_name = "./results/DistilBERT-combined"
pred_combined_model_val_data = evaluate_fine_tuned_llm(model_name, eval_dataset=dataset["test"])

def analyze_predictions(dataset, predictions):
    data_copy = dataset["test"].to_pandas()

    # Add predictions to scraped data
    data_copy["predicted_label"] = predictions

    # Print predictions by source
    print("\nPredictions by Source:")
    source_predictions = (
        data_copy.groupby("source")["predicted_label"].value_counts().unstack()
    )
    source_predictions.columns = ["Predicted Real", "Predicted Fake"]
    print(source_predictions)

    # Calculate percentage of fake news by source
    fake_percentages = data_copy.groupby("source")["predicted_label"].mean() * 100
    print("\nPercentage of Fake News by Source:")
    print(fake_percentages.sort_values(ascending=False))

analyze_predictions(dataset, pred_combined_model_val_data)

%%time

model_name = "./results/DistilBERT-original"
pred_original_model_val_data = evaluate_fine_tuned_llm(model_name, eval_dataset=dataset["test"])

analyze_predictions(dataset, pred_original_model_val_data)
